Fixes for .NET Core 3.1 update
