+++ ChustaSoft.Tools.Authorization/ChustaSoft.Tools.Authorization.csproj
@@ -21,6 +21,8 @@
     
     <PackageReference Include="ChustaSoft.Common" Version="1.2.0" />
     
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="3.1.2" />
+    
     <PackageReference Include="Microsoft.AspNetCore.Identity" Version="2.2.0" />
     <PackageReference Include="Microsoft.EntityFrameworkCore" Version="3.1.2" />
     <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="3.1.2" />
